refactor(BrainUsers)
