+++ frontend/lib/components/NavBar/components/NavItems/components/BrainsDropDown/components/BrainActions/components/ShareBrain/components/index.ts
@@ -1,2 +1,2 @@
-export * from "./BrainUser";
+export * from "./BrainUsers/BrainUser";
 export * from "./UserToInvite";
