+++ frontend/lib/components/NavBar/components/NavItems/components/BrainsDropDown/components/BrainActions/components/ShareBrain/hooks/useShareBrain.ts
@@ -1,9 +1,7 @@
 /* eslint-disable max-lines */
-import { useEffect, useState } from "react";
+import { useState } from "react";
 
-import { Subscription } from "@/lib/api/brain/brain";
 import { useBrainApi } from "@/lib/api/brain/useBrainApi";
-import { useSupabase } from "@/lib/context/SupabaseProvider";
 import { useToast } from "@/lib/hooks";
 
 import { BrainRoleAssignation } from "../../../types";
@@ -11,8 +9,6 @@ import { generateBrainAssignation } from "../utils/generateBrainAssignation";
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export const useShareBrain = (brainId: string) => {
-  const [brainUsers, setBrainUsers] = useState<Subscription[]>([]);
-  const [isFetchingBrainUsers, setFetchingBrainUsers] = useState(false);
   const [sendingInvitation, setSendingInvitation] = useState(false);
   const [isShareModalOpen, setIsShareModalOpen] = useState(false);
   const [roleAssignations, setRoleAssignation] = useState<
@@ -23,8 +19,7 @@ export const useShareBrain = (brainId: string) => {
   const brainShareLink = `${baseUrl}/invitation/${brainId}`;
 
   const { publish } = useToast();
-  const { addBrainSubscriptions, getBrainUsers } = useBrainApi();
-  const { session } = useSupabase();
+  const { addBrainSubscriptions } = useBrainApi();
 
   const handleCopyInvitationLink = async () => {
     await navigator.clipboard.writeText(brainShareLink);
@@ -95,26 +90,6 @@ export const useShareBrain = (brainId: string) => {
     setRoleAssignation([...roleAssignations, generateBrainAssignation()]);
   };
 
-  const fetchBrainUsers = async () => {
-    // Optimistic update
-    setFetchingBrainUsers(brainUsers.length === 0);
-    try {
-      const users = await getBrainUsers(brainId);
-      setBrainUsers(users.filter(({ email }) => email !== session?.user.email));
-    } catch {
-      publish({
-        variant: "danger",
-        text: "An error occurred while fetching brain users",
-      });
-    } finally {
-      setFetchingBrainUsers(false);
-    }
-  };
-
-  useEffect(() => {
-    void fetchBrainUsers();
-  }, []);
-
   return {
     roleAssignations,
     brainShareLink,
@@ -126,8 +101,5 @@ export const useShareBrain = (brainId: string) => {
     sendingInvitation,
     setIsShareModalOpen,
     isShareModalOpen,
-    brainUsers,
-    fetchBrainUsers,
-    isFetchingBrainUsers,
   };
 };
